move vlr compression implementation into separate .cpp
